+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsMediaPlaylist.java
@@ -391,8 +391,13 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
 
   /** The type of the playlist. See {@link PlaylistType}. */
   @PlaylistType public final int playlistType;
-  /** The start offset in microseconds, as defined by #EXT-X-START. */
+  /**
+   * The start offset in microseconds, as defined by #EXT-X-START, or {@link C#TIME_UNSET} if
+   * undefined.
+   */
   public final long startOffsetUs;
+  /** Whether the start position should be precise, as defined by #EXT-X-START. */
+  public final boolean preciseStart;
   /**
    * If {@link #hasProgramDateTime} is true, contains the datetime as microseconds since epoch.
    * Otherwise, contains the aggregated duration of removed segments up to this snapshot of the
@@ -467,6 +472,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
       String baseUri,
       List<String> tags,
       long startOffsetUs,
+      boolean preciseStart,
       long startTimeUs,
       boolean hasDiscontinuitySequence,
       int discontinuitySequence,
@@ -485,6 +491,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
     super(baseUri, tags, hasIndependentSegments);
     this.playlistType = playlistType;
     this.startTimeUs = startTimeUs;
+    this.preciseStart = preciseStart;
     this.hasDiscontinuitySequence = hasDiscontinuitySequence;
     this.discontinuitySequence = discontinuitySequence;
     this.mediaSequence = mediaSequence;
@@ -562,6 +569,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
         baseUri,
         tags,
         startOffsetUs,
+        preciseStart,
         startTimeUs,
         /* hasDiscontinuitySequence= */ true,
         discontinuitySequence,
@@ -592,6 +600,7 @@ public final class HlsMediaPlaylist extends HlsPlaylist {
         baseUri,
         tags,
         startOffsetUs,
+        preciseStart,
         startTimeUs,
         hasDiscontinuitySequence,
         discontinuitySequence,